requirements.txt
numpy>=1.19.5
librosa>=0.8.1
tensorflow>=2.8.0
soundfile>=0.10.3
scikit-learn>=1.0.2
matplotlib>=3.5.1
pandas>=1.3.5
tqdm>=4.62.3
pyyaml>=6.0
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
python-socketio>=5.4.0
eventlet>=0.33.0
